+++ bmmodel/guardian/guardian.go
@@ -17,9 +17,9 @@ type BmGuardian struct {
 	Name     string `json:"name" bson:"name"`
 	Nickname string `json:"nickname" bson:"nickname"`
 	Icon     string `json:"icon" bson:"icon"`
-	Dob      int64  `json:"dob" bson:"dob"`
-	Gender   int64  `json:"gender" bson:"gender"`
-	RegDate  int64  `json:"reg_date" bson:"reg_date"`
+	Dob      float64  `json:"dob" bson:"dob"`
+	Gender   float64  `json:"gender" bson:"gender"`
+	RegDate  float64  `json:"reg_date" bson:"reg_date"`
 	Contact  string `json:"contact" bson:"contact"`
 	WeChat   string `json:"wechat" bson:"wechat"`
